backend/src/analytics/requirements.txt
fastapi
uvicorn
supabase
python-consul
prometheus-fastapi-instrumentator
python-dotenv